Move container parser states to another module
